+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Release 5.5.4:
  - Token status:
    - Add considerations for separating the semantics "no token status mechanism is defined" from "evaluating token status failed"
    - Provide revocation status to verifier
+ - DCQL:
+   - Parse new format of claim query in OpenID4VP Draft 28
 
 Release 5.5.3:
  - Fix DCQL Query serialization/deserialization in `AuthenticationRequestParameters`